Make strata module respond to build methods
The strata module itself now has all build/route methods. This should
significantly simplify the API for newcomers. Also added route methods
back to build middleware. The caveat is that calls to the route methods
are not interleaved with calls to map and/or use. Instead, if any of
them are used the resulting route middleware is simply used as that
build's fall-through app.