Fix HEAD so w/o length returns chunked
Fix the issue where HEAD without a Content-Length should return
Transfer-Encoding chunked to match the corresponding request from
a GET.

http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13

Make sure that Transfer-Encoding is only set for streams.

I had to split the body = '' set into two places since
we need to know whether it was a string or stream in the
Transfer-Encoding check.